+++ lib/modules/forms/forms_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: forms
-// Updated at: 2026-08-19 09:38:24
-// Commit: test(forms): implement widget integration tests for forms
+// Updated at: 2026-08-19 14:39:14
+// Commit: feat(forms): implement shimmer loading skeleton in forms
 
 class formsComponentState {
-  final String commitRef = '3732';
-  final String lastMessage = 'test(forms): implement widget integration tests for forms';
-  final DateTime timestamp = DateTime.parse('2026-08-19T09:38:24+05:00');
+  final String commitRef = '3740';
+  final String lastMessage = 'feat(forms): implement shimmer loading skeleton in forms';
+  final DateTime timestamp = DateTime.parse('2026-08-19T14:39:14+05:00');
 }